requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
filelock==3.13.1
fsspec==2024.6.1
huggingface-hub==0.33.1
idna==3.10
Jinja2==3.1.4
joblib==1.5.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==25.0
pandas==2.3.0
pillow==11.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
setuptools==70.2.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.53.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
